+++ packages/sdks/core-js-sdk/src/sdk/types.ts
@@ -103,6 +103,7 @@ export type TemplateOptions = Record<string, string>; // for providing messaging
 export type LoginOptions = {
   stepup?: boolean;
   mfa?: boolean;
+  revokeOtherSessions?: boolean;
   customClaims?: Record<string, any>;
   templateId?: string;
   templateOptions?: TemplateOptions;
@@ -325,6 +326,7 @@ export type Options = {
   samlIdpStateId?: string;
   samlIdpUsername?: string;
   ssoAppId?: string;
+  thirdPartyAppId?: string;
   oidcLoginHint?: string;
   abTestingKey?: number;
   startOptionsVersion?: number;
@@ -333,6 +335,8 @@ export type Options = {
   oidcPrompt?: string;
   oidcErrorRedirectUri?: string;
   nativeOptions?: NativeOptions;
+  thirdPartyAppStateId?: string;
+  applicationScopes?: string; // Relevant for sso application and third party application
 };
 
 export type ResponseData = Record<string, any>;