+++ packages/sdks/core-js-sdk/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.33.0](https://github.com/descope/descope-js/compare/core-js-sdk-2.32.0...core-js-sdk-2.33.0) (2024-12-08)
+
+
+### Features
+
+* Add thirdPartyAppStateId and scopes parameters for third party application ([#856](https://github.com/descope/descope-js/issues/856)) ([fa95d30](https://github.com/descope/descope-js/commit/fa95d30810599cf4a198b09b96b36b7e4c284464))
+
+## [2.32.0](https://github.com/descope/descope-js/compare/core-js-sdk-2.31.0...core-js-sdk-2.32.0) (2024-12-04)
+
+
+### Features
+
+* Css vars ([#853](https://github.com/descope/descope-js/issues/853)) ([a49be2b](https://github.com/descope/descope-js/commit/a49be2b67b7eb8e3535647a94960f59396c70a0b))
+
+## [2.31.0](https://github.com/descope/descope-js/compare/core-js-sdk-2.30.1...core-js-sdk-2.31.0) (2024-11-16)
+
+
+### Features
+
+* Add the revokeOtherSessions attribute in LoginOptions RELEASE ([#848](https://github.com/descope/descope-js/issues/848)) ([fca695f](https://github.com/descope/descope-js/commit/fca695fa4ab079791c969590694912a5b2daf74f))
+
+## [2.30.1](https://github.com/descope/descope-js/compare/core-js-sdk-2.30.0...core-js-sdk-2.30.1) (2024-11-14)
+
 ## [2.30.0](https://github.com/descope/descope-js/compare/core-js-sdk-2.29.1...core-js-sdk-2.30.0) (2024-11-13)